+++ packages/api-reference/src/features/api-client-modal/ApiClientModal.vue
@@ -59,13 +59,18 @@ onMounted(() => {
 // Ensure we have a document when doing the initial import
 watchDebounced(
   () => dereferencedDocument,
-  (newDocument) => {
+  (newDocument, oldDocument) => {
     if (!newDocument) {
       return
     }
 
+    // Ensure the document is different
+    const diff = microdiff(newDocument, oldDocument || {})
+    if (!diff?.length) {
+      return
+    }
+
     // If we already have a collection, remove the store
-    // TODO: add @redis diffing here... or just upgrade to the new store
     if (activeEntities.activeCollection.value) {
       client.value?.resetStore()
     }