+++ integrations/docker/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @scalarapi/docker-api-reference
 
+## 0.2.21
+
+### Patch Changes
+
+- Updated dependencies [83625c6]
+  - @scalar/api-reference@1.34.2
+
+## 0.2.20
+
+### Patch Changes
+
+- Updated dependencies [4d05839]
+  - @scalar/api-reference@1.34.1
+
 ## 0.2.19
 
 ### Patch Changes